#5 #10 LoadUser.nickCheck 추가 및 @AuthenticationPrincipal 제거

+++ src/main/java/com/example/backend/board/repository/BoardRepository.java
@@ -3,5 +3,5 @@ package com.example.backend.board.repository;
 import com.example.backend.board.domain.Board;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BoardRepository extends JpaRepository<Board, Integer> {
+public interface BoardRepository extends JpaRepository<Board, Long> {
 }
